+++ CoreCms.Net.Model/CoreCms.Net.Model.csproj
@@ -1,11 +1,11 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net7.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="sqlSugarCore" Version="5.0.4.2" />
+    <PackageReference Include="sqlSugarCore" Version="5.1.3.50" />
   </ItemGroup>
 
 </Project>
