+++ CoreCms.Net.Services/Bill/CoreCmsBillDeliveryServices.cs
@@ -24,7 +24,6 @@ using CoreCms.Net.Model.ViewModels.UI;
 using CoreCms.Net.Model.ViewModels.DTO;
 using CoreCms.Net.Utility.Helper;
 using Flurl.Http;
-using Google.Protobuf.WellKnownTypes;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Newtonsoft.Json;
