+++ README.md
@@ -38,7 +38,7 @@ QQ群：70494974
 </div>
 
 #### 💐 核心商城（CoreShop）介绍
-> 核心商城系统`（CoreShop）` 是基于 Asp.Net 5.0、Uni-App开发、支持可视化布局的小程序商城系统；前后端分离，支持分布式部署，跨平台运行；拥有分销、代理、团购秒杀、接龙、拼团、直播、优惠券、自定义表单等众多营销功能，拥有完整SKU、下单、售后、物流流程，支持可视化自定义首页模块布局效果。    
+> 核心商城系统`（CoreShop）` 是基于 Asp.Net 7.0、Uni-App开发、支持可视化布局的小程序商城系统；前后端分离，支持分布式部署，跨平台运行；拥有分销、代理、团购秒杀、接龙、拼团、直播、优惠券、自定义表单等众多营销功能，拥有完整SKU、下单、售后、物流流程，支持可视化自定义首页模块布局效果。    
 
 > 支持一套代码编译发布微信小程序版、H5版、Android版、iOS版、支付宝小程序版、百度小程序版、字节跳动小程序版、QQ小程序版、快应用、360小程序共10个平台。
